0x05-pointers_arrays_strings/100-atoi.c

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@ int _atoi(char *s)
 {
 	int sign = 1;
 
-	long long num = 0;
+	long int num = 0;
 
 	int i = 0;
